requirements.txt
# Unified Trading System - Dependencies
# Phase 1: Foundation
# Last Updated: March 4, 2026

# ================================
# Core Data Processing (ALL)
# ================================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
pyarrow>=12.0.0

# ================================
# IBKR Connection
# ================================
ib-insync==0.9.86

# ================================
# Market Data Sources
# ================================
yfinance>=0.2.0              # Yahoo Finance (free)

# ================================
# Technical Analysis
# ================================
ta>=0.10.0                   # Simple TA library
# ta-lib==0.4.28             # Uncomment for advanced TA (requires system install)

# ================================
# Web Scraping (Finviz)
# ================================
selenium>=4.0.0
webdriver-manager>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# ================================
# Configuration & Validation
# ================================
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.0.0

# ================================
# Utilities
# ================================
python-dateutil>=2.8.2
pytz>=2024.0
tenacity>=8.2.0              # Retry logic

# ================================
# Logging & Monitoring
# ================================
python-json-logger>=2.0.7
watchdog>=3.0.0              # File system monitoring

# ================================
# Streamlit UI (Primary)
# ================================
streamlit>=1.31.0
plotly>=5.18.0               # Interactive charts
streamlit-autorefresh>=1.0.0 # Auto-refresh for live data
streamlit-aggrid>=0.3.4      # Advanced tables

# ================================
# Alternative APIs (Optional)
# ================================
fastapi>=0.109.0             # For REST APIs
uvicorn[standard]>=0.27.0    # ASGI server
websockets>=12.0             # WebSocket support

# ================================
# Database
# ================================
sqlalchemy>=2.0.0
aiosqlite>=0.19.0            # Async SQLite

# ================================
# AI Features (Optional)
# ================================
openai>=1.9.0                # For LLM trade narratives

# ================================
# Email Notifications (Optional)
# ================================
sendgrid>=6.11.0             # SendGrid API

# ================================
# Testing
# ================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# ================================
# Development Tools
# ================================
black>=24.0.0                # Code formatter
flake8>=7.0.0                # Linter
mypy>=1.8.0                  # Type checker
